Consistencia lunes
Consitencia de: editar asesor, eliminar asesor, para ello se paso las clases de user a consultant, además se agrego el rol al seeder de usuario

+++ app/Http/Controllers/ConsultantController.php
@@ -3,21 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
 use App\Models\Consultant as ConsultantModel;
 use App\Http\Requests\Consultant;
-
-use Illuminate\Database\Eloquent\ModelNotFoundException;
-
-use Illuminate\Support\Facades\Hash;
+use App\User as UserModel;
 
 
 class ConsultantController extends Controller
 {
     public function index()
     {
-        $consultants = ConsultantModel::all();
+        $consultants = UserModel::all();
 
-        return view('consultants.index', [
+        return view('consultants.manage', [
             'consultants' => $consultants,
         ]);
     }
@@ -31,10 +31,10 @@ class ConsultantController extends Controller
     {
         $data = $request->validated();
         $newConsultant = new ConsultantModel;
-        $newConsultant-> password = Hash::make($request->input('password'));
+        $newConsultant->password = Hash::make($request->input('password'));
         $newConsultant->fill($data);
         $newConsultant->save();
-        return back()->with('mensaje','El asesor fue registrado exitosamente');
+        return back()->with('mensaje', 'El asesor fue registrado exitosamente');
     }
 
     public function show($id)
@@ -43,14 +43,12 @@ class ConsultantController extends Controller
     }
 
     public function edit($id)
-    {   
+    {
         try {
-            $consultant= ConsultantModel::findOrFail($id);
+            $consultant = UserModel::findOrFail($id);
+        } catch (ModelNotFoundException $e) {
+            return view('consultants')->with('mensajeError', 'No se pudo encontrar el asesor a editar');
         }
-        catch (ModelNotFoundException $e){
-            return redirect('consultants')->with('mensajeError','No se pudo encontrar el asesor a editar');  
-        }
-
         return view('consultants.edit', [
             'consultant' => $consultant,
         ]);
@@ -58,28 +56,58 @@ class ConsultantController extends Controller
 
     public function update(Consultant $request, $id)
     {
-        $data = $request->validated();
-       
         try {
-            $consultantUpdate= ConsultantModel::findOrFail($id);
+            $userUpdate = UserModel::findOrFail($id);
+        } catch (ModelNotFoundException $e) {
+            return back()->with('mensajeError', 'No se pudo encontrar el asesor a editar');
         }
-        catch (ModelNotFoundException $e){
-            return redirect('consultants')->with('mensajeError','No se pudo encontrar el asesor a editar');  
-        }
-        $consultantUpdate->fill($data);
-        $consultantUpdate->save();
-        return redirect('consultants')->with('mensaje','El asesor fue actualizado exitosamente');
+
+        Validator::make($request->all(), [
+            'identification' => 'required|min:0|integer|digits_between:5,15|unique:users' . ($id ? ",identification,$id" : ''),
+            'name' => 'required|between:3,30',
+            'email' => 'required|email|unique:users' . ($id ? ",email,$id" : ''),
+            'password' => 'required|min:6'
+        ], [
+            'identification.required' => 'La identificación es obligatoria',
+            'identification.min' => 'La identificación no puede ser negativa',
+            'identification.integer' => 'La identificación debe ser un número entero',
+            'identification.digits_between' => 'La identificación debe contener entre 5 y 15 digítos',
+            'identification.unique' => 'La identificación ya está en uso',
+            'name.required' => 'El nombre es obligatorio',
+            -'name.between' => 'El nombre debe contener entre 3 y 30 caracteres',
+            'email.required' => 'El correo electrónico es obligatorio',
+            'email.email' => 'El correo electrónico es incorrecto',
+            'email.unique' => 'El correo electrónico ya está en uso',
+            'password.required' => 'La contraseña es obligatoria',
+            'password.min' => 'La contraseña debe tener al menos 6 caracteres',
+        ])->validate();
+
+        $userUpdate->update([
+            'identification' => $request['identification'],
+            'name' => $request['name'],
+            'email' => $request['email'],
+            'password' => Hash::make($request['password']),
+        ]);
+
+        return redirect('consultants')->with('mensaje', 'El usuario fue actualizado exitosamente');
     }
 
     public function destroy($id)
     {
         try {
-        $consultantDelete = ConsultantModel::findOrFail($id);
+            $userDelete = UserModel::findOrFail($id);          
+        } catch (ModelNotFoundException $e) {
+            return redirect('consultants')->with('mensajeError', 'No se pudo encontrar el asesor a eliminar');
         }
-        catch (ModelNotFoundException $e){
-            return redirect('consultants')->with('mensajeError','No se pudo encontrar el asesor a eliminar');  
+
+        $consultantDelete = ConsultantModel::where('user_id', $id);
+        if ($consultantDelete == null) {
+            return redirect('consultants')->with('mensajeError', 'No se pudo encontrar el asesor a eliminar');
         }
+
         $consultantDelete->delete();
-        return redirect('consultants')->with('mensaje','Asesor Eliminado exitosamente');        
+        $userDelete->delete();
+
+        return redirect('consultants')->with('mensaje', 'Asesor Eliminado exitosamente');
     }
 }
